% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOSTAT625ttest-package.R
\docType{package}
\name{BIOSTAT625ttest-package}
\alias{BIOSTAT625ttest}
\alias{BIOSTAT625ttest-package}
\title{BIOSTAT625ttest: A one-sample, Student's t-test}
\description{
This test package has a simple function to run one-sample t-test.
}
\keyword{internal}
