% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{length_cpp}
\alias{length_cpp}
\title{Finds the length of a vector (or array)}
\usage{
length_cpp(x)
}
\arguments{
\item{x}{Numeric vector in R}
}
\description{
Finds the length of a vector (or array)
}
