% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sd_cpp}
\alias{sd_cpp}
\title{Finds the standard deviation of a vector (or array)}
\usage{
sd_cpp(x)
}
\arguments{
\item{x}{Numeric vector in R}
}
\description{
Finds the standard deviation of a vector (or array)
}
