% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOSTAT625ttest-package.R
\docType{package}
\name{BIOSTAT625ttest-package}
\alias{BIOSTAT625ttest}
\alias{BIOSTAT625ttest-package}
\title{BIOSTAT625ttest: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
